Update wording:
This sort of revert the changes made in https://github.com/rwz/cant-parse-emails/pull/20 which is, in my understanding, blurring the lines between what is standard and what is not. I think the wording prior to that PR were better as they put more emphasis on the fact that supporting `+` in the local part is standard. Mail clients should redirect to the proper mailbox, even if some don’t.